move renderstates to gl
